kill struct xfs_mount_args
No need to parse the mount option into a structure before applying it
to struct xfs_mount.

The content of xfs_start_flags gets merged into xfs_parseargs.  Calls
inbetween don't care and can use mount members instead of the args
struct.

This patch uncovered that the mount option for shared filesystems wasn't
ever exposed on Linux.  The code to handle it is #if 0'ed in this patch
pending a decision on this feature.  I'll send a writeup about it to
the list soon.


Signed-off-by: Christoph Hellwig <hch@lst.de>
Merge of xfs-linux-melb:xfs-kern:32371a by kenmcd.

  kill struct xfs_mount_args