fix biosize option
iosizelog shouldn't be the same as iosize but the logarithm of it.  Then
again the current biosize option doesn't make much sense to me as it
doesn't set the preferred I/O size as mentioned in the comment next to
it but rather the allocation size and thus is identical to the allocsize
option (except for the missing logarithm).  It's also not documented in
Documentation/filesystems/xfs.txt or the mount manpage.


Signed-off-by: Christoph Hellwig <hch@lst.de>
Merge of xfs-linux-melb:xfs-kern:32373a by kenmcd.

  fix biosize option