stop using xfs_itobp in xfs_bulkstat
xfs_bulkstat only wants the dinode, offset and buffer from a given
inode number.  Instead of using xfs_itobp on a fake inode which is
complicated and currently leads to leaks of the security data just use
xfs_inotobp which is designed to do exactly the kind of lookup
xfs_bulkstat wants.  The only thing that's missing in xfs_inotobp is
a flags paramter that let's us pass down XFS_IMAP_BULKSTAT, but that can
easily added.


Signed-off-by: Christoph Hellwig <hch@lst.de>
Merge of xfs-linux-melb:xfs-kern:32397a by kenmcd.

  stop using xfs_itobp in xfs_bulkstat